+++ src/Drupal/Config/BuildConfiguration.php
@@ -33,7 +33,8 @@ class BuildConfiguration implements ConfigurationInterface
                     ->isRequired()
                     ->cannotBeEmpty()
                     ->validate()
-                    ->ifTrue(function ($alias) { return substr($alias, 0, 1) !== "@";
+                    ->ifTrue(function ($alias) {
+                        return substr($alias, 0, 1) !== "@";
                     })
                         ->thenInvalid("Drush aliases must be specified with a leading @")
                     ->end()
@@ -86,6 +87,17 @@ class BuildConfiguration implements ConfigurationInterface
                 // Database details.
                 ->arrayNode("database")
                     ->children()
+                        ->enumNode("driver")
+                            ->values(array("mysql", "pgsql", "sqlite"))
+                            ->defaultValue("mysql")
+                        ->end()
+                        ->scalarNode("host")
+                            ->defaultValue("localhost")
+                        ->end()
+                        ->integerNode("port")
+                            ->min(1)
+                            ->defaultValue(3306)
+                        ->end()
                         ->scalarNode("db_name")
                             ->isRequired()
                             ->cannotBeEmpty()
@@ -102,7 +114,9 @@ class BuildConfiguration implements ConfigurationInterface
                 ->end()
 
                 // Pre steps.
-                ->append($this->addPreOrPostSteps("pre"))
+                ->arrayNode("pre")
+                    ->prototype("scalar")->end()
+                ->end()
 
                 // Enable Features modules.
                 ->arrayNode("features")
@@ -113,46 +127,13 @@ class BuildConfiguration implements ConfigurationInterface
                 ->append($this->addMigrateSection())
 
                 // Post steps.
-                ->append($this->addPreOrPostSteps("post"))
+                ->arrayNode("post")
+                    ->prototype("scalar")->end()
+                ->end()
             ->end()
         ;
 
         return $treeBuilder;
-    }
-
-    /**
-     * Build and return a 'pre' or 'post' step configuration node.
-     *
-     * @param string $step
-     *   The step to build, either 'pre' or 'post'.
-     *
-     * @return ArrayNodeDefinition|NodeDefinition
-     *   The built node definition.
-     *
-     * @throws \Exception
-     */
-    public function addPreOrPostSteps($step)
-    {
-        if ($step != "pre" && $step != 'post') {
-            throw new \Exception("$step is not a valid build step, must be 'pre' or 'post'.");
-        }
-
-        $builder = new TreeBuilder();
-        $node = $builder->root($step);
-
-        $node
-            ->canBeEnabled()
-            ->children()
-                ->arrayNode("modules")
-                    ->prototype("scalar")->end()
-                ->end()
-                ->arrayNode("commands")
-                    ->prototype("scalar")->end()
-                ->end()
-            ->end()
-        ;
-
-        return $node;
     }
 
     /**