+++ src/Drupal/DrupalBuild.php
@@ -30,6 +30,40 @@ class DrupalBuild
     const BUILD_CONFIG_FILE = "drupal.build.yml";
 
     /**
+     * The format of the database URL passed to drush site-install
+     */
+    const DRUPAL_DB_URL_SYNTAX = "%s://%s:%s@%s:%d/%s";
+
+    /**
+     * @var array
+     *   A list of Drupal's hidden files (to remove).
+     */
+    public static $drupalHiddenFiles = array('.htaccess', '.gitignore');
+
+    /**
+     * @var array
+     *   List of file patterns to recursively remove during cleanup.
+     */
+    public static $unwantedFilesPatterns = array(
+        '*txt',
+        'install.php',
+        'scripts',
+        'web.config',
+    );
+
+    /**
+     * @var string
+     *   The definition of a line/pattern in $sites.php
+     */
+    public static $sitesFileLinePattern = "\$sites['%s'] = '%s';";
+
+    /**
+     * @var string
+     *   Drupal's default theme.
+     */
+    public static $drupalDefaultTheme = 'bartik';
+
+    /**
      * @var \Robo\DrupalRunner
      *   Stores the calling DrupalRunner instance.
      */
@@ -188,18 +222,20 @@ class DrupalBuild
 if (file_exists(conf_path() . '/settings.$env.php')) {
   include_once 'settings.$env.php';
 }
+
 EOS;
 
-        // Write the inclusion of environment specific configuration to main settings.php file.
+        // If the string is NOT already present in the file (e.g. from a previous build), write the inclusion of
+        // environment specific configuration to main settings.php file.
         $settingsFilePath = "sites/{$this->config['sites_subdir']}/settings.php";
-        $this->taskExec("chmod u+w {$this->path($settingsFilePath)}")->run();
-
-        $this->taskWriteToFile($this->path($settingsFilePath))
-            ->text($envSettings)
-            ->append()
-            ->run();
-
-        $this->taskExec("chmod u-w {$this->path($settingsFilePath)}")->run();
+        if (strpos(file_get_contents($this->path($settingsFilePath)), $envSettings) === false) {
+            $this->taskExec("chmod u+w {$this->path($settingsFilePath)}")->run();
+            $this->taskWriteToFile($this->path($settingsFilePath))
+                ->text($envSettings)
+                ->append()
+                ->run();
+            $this->taskExec("chmod u-w {$this->path($settingsFilePath)}")->run();
+        }
     }
 
     /**